added authorisation

+++ src/services/animalType.service.ts
@@ -1,3 +1,4 @@
+import { Role } from '@prisma/client';
 import { prisma } from '../config/prisma';
 import { HttpException } from '../exceptions/httpException';
 
@@ -24,9 +25,12 @@ const getAnimalType = async (id: string) => {
   return animalType;
 };
 
-const createAnimalType = async (data: { name: string; description: string }) => {
+const createAnimalType = async (data: { name: string; description: string }, role: string) => {
   if (!data.name || !data.description) {
     throw new HttpException(400, 'Bad request!');
+  }
+  if (role != Role.ADMIN && role != Role.SUPER_ADMIN) {
+    throw new HttpException(403, 'Forbidden.');
   }
   const animalType = await prisma.animalType.create({
     data: { name: data.name, description: data.description },
@@ -34,10 +38,17 @@ const createAnimalType = async (data: { name: string; description: string }) => 
   return animalType;
 };
 
-const updateAnimalType = async (data: { name: string; description: string }, id: string) => {
+const updateAnimalType = async (
+  data: { name: string; description: string },
+  id: string,
+  role: string
+) => {
   const temp = await prisma.animalType.findUnique({ where: { id } });
   if (!temp) {
     throw new HttpException(404, 'Not found!');
+  }
+  if (role != Role.ADMIN && role != Role.SUPER_ADMIN) {
+    throw new HttpException(403, 'Forbidden.');
   }
   const animalType = await prisma.animalType.update({
     where: { id: id },
@@ -46,10 +57,13 @@ const updateAnimalType = async (data: { name: string; description: string }, id:
   return animalType;
 };
 
-const deleteAnimalType = async (id: string) => {
+const deleteAnimalType = async (id: string, role: string) => {
   const temp = await prisma.animalType.findUnique({ where: { id } });
   if (!temp) {
     throw new HttpException(404, 'Not found!');
+  }
+  if (role != Role.ADMIN && role != Role.SUPER_ADMIN) {
+    throw new HttpException(403, 'Forbidden.');
   }
   const animalType = await prisma.animalType.delete({
     where: { id: id },
