+++ src/services/topic.service.ts
@@ -1,3 +1,4 @@
+import { Role } from '@prisma/client';
 import { prisma } from '../config/prisma';
 import { HttpException } from '../exceptions/httpException';
 
@@ -42,9 +43,16 @@ const getAnimalTypeTopics = async (animalTypeId: string) => {
   return animalTypeTopics;
 };
 
-const createTopic = async (animalTypeId: string, data: { title: string; description: string }) => {
+const createTopic = async (
+  animalTypeId: string,
+  data: { title: string; description: string },
+  role: string
+) => {
   if (!data.title || !data.description) {
     throw new HttpException(400, 'Bad request!');
+  }
+  if (role != Role.ADMIN && role != Role.SUPER_ADMIN) {
+    throw new HttpException(403, 'Forbidden.');
   }
   const topic = await prisma.topic.create({
     data: { title: data.title, description: data.description, animalTypeId: animalTypeId },
@@ -55,11 +63,15 @@ const createTopic = async (animalTypeId: string, data: { title: string; descript
 const updateTopic = async (
   animalTypeId: string,
   id: string,
-  data: { title: string; description: string }
+  data: { title: string; description: string },
+  role: string
 ) => {
   const tempTopic = await prisma.topic.findFirst({ where: { id: id, animalTypeId } });
   if (!tempTopic) {
     throw new HttpException(404, 'Not found!');
+  }
+  if (role != Role.ADMIN && role != Role.SUPER_ADMIN) {
+    throw new HttpException(403, 'Forbidden.');
   }
   const topic = await prisma.topic.update({
     where: { id: id },
@@ -68,10 +80,13 @@ const updateTopic = async (
   return topic;
 };
 
-const deleteTopic = async (animalTypeId: string, id: string) => {
+const deleteTopic = async (animalTypeId: string, id: string, role: string) => {
   const tempTopic = await prisma.topic.findFirst({ where: { id: id, animalTypeId } });
   if (!tempTopic) {
     throw new HttpException(404, 'Not found!');
+  }
+  if (role != Role.ADMIN && role != Role.SUPER_ADMIN) {
+    throw new HttpException(403, 'Forbidden.');
   }
   const topic = await prisma.topic.delete({
     where: { id: id },
